Add management commands
- ./manage.py show_state <scope>

Formats the state and prints to stdout

- ./manage.py regenerate_from_raw

Useful when developing the serializers, or adding a field to the models.

The request's raw JSON is saved on the State model's `raw` field.
Updating the model (e.g. with a migration) and backfilling any fields
can be triggered with this script without requiring another terraform run.